refactor to fix test
