+++ test/unit/runtime-config.env.test.ts
@@ -1,11 +1,6 @@
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 
-vi.mock("#nitro-internal-virtual/runtime-config", () => ({
-  runtimeConfig: {},
-}));
-
-const { applyEnv } =
-  await import("../../src/runtime/internal/runtime-config.ts");
+import { applyEnv } from "../../src/runtime/internal/runtime-config.utils.ts";
 
 describe("env utils", () => {
   describe("applyEnv", () => {
